Add support for razor pages and add razor pages sample
Allow using the [ValidateReCaptcha] on classes as well as on methods, because it
triggers a warning when used on a razor page handler. Also check if the request
has a form content type to not check the request on GET requests.

Also removed the old .NET Core 2.1 sample and renamed NetCore50 to Net50

Fixes #10